requirements.txt
absl-py==2.3.1
slither-analyzer==0.10.2
solc-select==1.0.4